requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2
django-filter==22.1
django-getenv==1.3.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
filelock==3.9.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.0
pipenv==2023.2.18
platformdirs==3.1.1
pluggy==1.0.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.2
python3-openid==3.2.0
pytz==2022.7.1
rave-python==1.2.16
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
